# server-python/openapi_server/models/new_pet.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class NewPet(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, tag=None):  # noqa: E501
        """NewPet - a model defined in OpenAPI

        :param name: The name of this NewPet.  # noqa: E501
        :type name: str
        :param tag: The tag of this NewPet.  # noqa: E501
        :type tag: str
        """
        self.openapi_types = {
            'name': str,
            'tag': str
        }

        self.attribute_map = {
            'name': 'name',
            'tag': 'tag'
        }

        self._name = name
        self._tag = tag

    @classmethod
    def from_dict(cls, dikt) -> 'NewPet':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The NewPet of this NewPet.  # noqa: E501
        :rtype: NewPet
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self):
        """Gets the name of this NewPet.


        :return: The name of this NewPet.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this NewPet.


        :param name: The name of this NewPet.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def tag(self):
        """Gets the tag of this NewPet.


        :return: The tag of this NewPet.
        :rtype: str
        """
        return self._tag

    @tag.setter
    def tag(self, tag):
        """Sets the tag of this NewPet.


        :param tag: The tag of this NewPet.
        :type tag: str
        """

        self._tag = tag